Empty menu is shown
When there are no namespaces then an empty 'API Documentation'
menu was shown in the navigation. This is not helpful and the menu
should not be shown if there is nothing to show.

In this commit I have added business rules to hide the menu if there
is nothing to display.

Related to https://github.com/phpDocumentor/phpDocumentor2/issues/1158