Commit Angular prototype
For version 2 of the Clean template we are in the process of switching
to an angular / jsonp powered version instead of generating each file
as static HTML.

The upside of this approach is that it speeds up the generating of a
piece of documentation by a significant factor but the downside is that
javascript is required to view the documentation and the minimum browser
version for IE is IE9.

In addition contributing to this template is more complex because it uses
Grunt, Bower, Compass and Angular as tools / frameworks.